adding FlatTreeLooperBase.py in preparation for multi loop
