+++ examples/all_examples.sh
@@ -6,5 +6,4 @@ python example_plot.py
 python example_stack.py
 python example_treeConverter.py
 python example_treeMaker.py
-python example_treeReader_extraBranches.py
 python example_treeReader.py
